--- Model/CoverableDesigner.php
@@ -1,23 +1,0 @@
-<?php declare(strict_types=1);
-/**
- * PHP version 7
- *
- * LICENSE: This source file is subject to copyright
- *
- * @package     App\Document
- * @author      Florian Ajir <florian@tag-walk.com>
- * @copyright   2016-2018 TAGWALK
- * @license     proprietary
- */
-
-namespace Tagwalk\ApiClientBundle\Model;
-
-use Tagwalk\ApiClientBundle\Model\Traits\Coverable;
-
-/**
- * Describe a Designer Document
- */
-class CoverableDesigner extends Designer
-{
-    use Coverable;
-}
