--- Serializer/Normalizer/CoverableDesignerNormalizer.php
@@ -1,71 +1,0 @@
-<?php
-/**
- * PHP version 7
- *
- * LICENSE: This source file is subject to copyright
- *
- * @author      Florian Ajir <florian@tag-walk.com>
- * @copyright   2016-2019 TAGWALK
- * @license     proprietary
- */
-
-namespace Tagwalk\ApiClientBundle\Serializer\Normalizer;
-
-use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
-use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-use Tagwalk\ApiClientBundle\Model\CoverableDesigner;
-use Tagwalk\ApiClientBundle\Model\File;
-
-/**
- * Normalizer for CoverableDesigner instances
- *
- * @extends DocumentNormalizer
- */
-class CoverableDesignerNormalizer extends DocumentNormalizer implements NormalizerInterface
-{
-    /**
-     * @var FileNormalizer
-     */
-    private $fileNormalizer;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function __construct(
-        NameConverterInterface $nameConverter = null,
-        PropertyAccessorInterface $propertyAccessor = null,
-        FileNormalizer $fileNormalizer
-    ) {
-        parent::__construct($nameConverter, $propertyAccessor);
-        $this->fileNormalizer = $fileNormalizer;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsNormalization($data, $format = null)
-    {
-        return $data instanceof CoverableDesigner;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsDenormalization($data, $type, $format = null)
-    {
-        return $type === CoverableDesigner::class;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function denormalize($data, $class, $format = null, array $context = [])
-    {
-        if (false === empty($data['cover'])) {
-            $data['cover'] = $this->fileNormalizer->denormalize($data['cover'], File::class);
-        }
-
-        return parent::denormalize($data, $class, $format, $context);
-    }
-}
